[bench] Support Android
